refactor: 모집일 관련 상수 데이터 리팩토링
