+++ src/components/home/HomeRecruitSection.tsx
@@ -2,7 +2,8 @@ import { useEffect, useState } from 'react'
 import { motion, AnimatePresence } from 'motion/react'
 import { Button1, CountdownCard } from '@/components/common'
 
-import { RECRUIT_SECTION_DATA, RECRUIT_TARGET_DATE } from '@/constants/home/recruitSection'
+import { RECRUIT_TARGET_DATE } from '@/constants/recruit'
+import { RECRUIT_SECTION_DATA } from '@/constants/home/recruitSection'
 import { useHomeRecruitAnimation } from '@/hooks/home/useHomeRecruitAnimation'
 
 type TimeLeft = { days: number; hours: number; minutes: number; seconds: number }
