+++ src/constants/home/recruitSection.ts
@@ -1,6 +1,3 @@
-/** 모집 마감 카운트다운 타겟 날짜 (KST) */
-export const RECRUIT_TARGET_DATE = new Date('2026-05-01T00:00:00+09:00')
-
 export const RECRUIT_SECTION_DATA = {
   title: 'Join to CMC',
   buttonHref: 'https://cmc.or.kr',
