--- src/constants/project/tab.ts
@@ -1,9 +1,0 @@
-export const PROJECT_TAB_ITEMS = [
-  { id: 'all', label: '전체' },
-  { id: '18', label: '18기' },
-  { id: '17', label: '17기' },
-  { id: '16', label: '16기' },
-  { id: '15', label: '15기' },
-  { id: '14', label: '14기' },
-  { id: '13', label: '13기' },
-] as const
